Fix redirect from HTTP to HTTPS
Fixed bug caused by redirect from HTTP to HTTPS. In this case the
protocol is updated, but simplecrawler attempts to crawl via the
previous redirected URLs port (previously 80) causing a 404/410 error
to be returned by some servers.

This was previously reported as follows:

https://github.com/cgiffard/node-simplecrawler/issues/145
https://github.com/cgiffard/node-simplecrawler/issues/208

Test cases are difficult to compose for this case because it requires a
working TLS/SSL server in the test code.