Fix race-condition causing incomplete crawls
Fixes #298

When the combination of number of items and fetch time was high enough
(when compared to the crawler's crawl interval), the crawler could stop
crawling before processing the discovery phase (in which the items would
be discovered from the latest fetched page).

This commit adds a new counter `_openDiscoveries` that works similarly
to the already existing logic for `_openRequests` and `_openListeners`.
With this, if a `crawl()` is reached while items are still being
discovered (given by `_openDiscoveries`, which is updated on fetch if
the crawl depth is suitable), the crawler will not stop.