Fix format in FS cache backend error constructor
The string formatting style used when constructing an error in the FS cache backend isn't actually supported in Node.js, and the error that it ultimately emits is not very helpful (i.e. `Cache storage of resource (%s) blocked by file: %s`). This switches the constructor to use template strings, which work in all versions of Node.js that SimpleCrawler supports.